gantt/ganttRenderer: add top x-axis for large Gantt charts
I am leveraging a Gantt chart to generate a years-long Roadmap from text.
This is for a large team, so it is quite large. A request I get from my
stakeholders is for a top labelling axis in parallel to the current bottom
one, so that the dates are labelled no matter where we are scrolled on
the screen.